# Predict *RateLimitError, return immediately without network call (9700x speedup when rate limit exceeded).