This (minor) optimization is possible because the client keeps track of the rate limits from previous API calls, and GitHub's rate limiting model is known in advance, and is predictable.

In situations where it's possible to predict that the network API call will result in a `*RateLimitError`, simply return it immediately instead of making the network API call to GitHub's servers. This is both faster for the client (by around 9700 times in my benchmarks) because it avoids the network round-trip latency, and better for GitHub's servers that won't need to needlessly reject network requests.

Note that this optimization only kicks in when the client is already exceeding the rate limit and it hasn't been reset yet. It has no effect during most normal usage when rate limit is not exceeded.

### Benchmarks

I wrote a small benchmark (it assumes you've already exhausted the GitHub API rate limit before you begin):

```Go
func BenchmarkClientDo_rateLimitError(b *testing.B) {
	gh := github.NewClient(nil)
	b.ResetTimer()
	for i := 0; i < b.N; i++ {
		_, _, err := gh.Repositories.Get("shurcooL", "vfsgen")
		if _, ok := err.(*github.RateLimitError); !ok {
			log.Println("no *github.RateLimitError error, but expected it")
		}
	}
}
```

Results before this PR (with a latency to api.github.com of around 85 ms):

```
$ go test -bench=. -benchmem
testing: warning: no tests to run
PASS
BenchmarkClientDo_rateLimitError-8	      20	  97057828 ns/op	   55922 B/op	     126 allocs/op
ok  		2.609s
```

Results after this PR:

```
go test -bench=. -benchmem
testing: warning: no tests to run
PASS
BenchmarkClientDo_rateLimitError-8	  200000	     10005 ns/op	    2763 B/op	      34 allocs/op
ok  		3.576s
```

In this limited scenario, it's a speedup by 9700 times.

Helps #152 and #153.