@shurcooL 2016-04-24
@gmlewis, I've addressed your current feedback, PTAL.