@shurcooL 2016-04-30
@willnorris Should I take the longer time to review to mean you're not as confident about this change and not sure how to respond to it, or have you simply not gotten around to it yet?

This change, at a high level, is very much in line with what you discussed/proposed in these two comments:

https://github.com/google/go-github/issues/152#issuecomment-70887913
https://github.com/google/go-github/pull/153#issuecomment-70884748

I think the high level design is solid.

About the implementation, I actually took a lot of time to come up with how to achieve the desired behavior, because I knew there were more than 1 sets of rate limits, depending on the route. But I think the final solution I came up with is pretty decent and works. But of course I welcome any potential suggestions for how to make the implementation even cleaner.

Let me know if I can do anything to help here.