@willnorris 2016-05-02
just a case of me being out of town when the PR came in, and then busy with other things.  Looking at it now.