@shurcooL 2016-05-04
Awesome, thanks! It's not every day one gets a chance to submit a PR with a 9700x perf improvement. :P